+++ odyssnet/__init__.py
@@ -2,7 +2,6 @@ __version__ = "2.2.0"
 
 from .core.network import OdyssNet
 from .training.trainer import OdyssNetTrainer
-from .training.chaos_optimizer import ChaosGrad
 from .utils.odyssstore import save_checkpoint, load_checkpoint, transplant_weights, get_checkpoint_info
 from .utils.neurogenesis import Neurogenesis
 from .utils.data import set_seed
@@ -11,7 +10,6 @@ from .utils.history import TrainingHistory
 __all__ = [
     'OdyssNet',
     'OdyssNetTrainer',
-    'ChaosGrad',
     'save_checkpoint',
     'load_checkpoint',
     'transplant_weights',